udp: add low latency socket poll support
Add upport for busy-polling on UDP sockets.
In __udp[46]_lib_rcv add a call to sk_mark_ll() to copy the napi_id
from the skb into the sk.
This is done at the earliest possible moment, right after we identify
which socket this skb is for.
In __skb_recv_datagram When there is no data and the user
tries to read we busy poll.

Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
Signed-off-by: Eliezer Tamir <eliezer.tamir@linux.intel.com>
Acked-by: Eric Dumazet <edumazet@google.com>
Tested-by: Willem de Bruijn <willemb@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Manual Change in udp.c as Hunk 2/2 fuzz (/* a return value > 0 means to resubmit the input, but -> ... value > 0 means to resubmit the input */)
https://github.com/torvalds/linux/commit/a5b50476f77a8fcc8055c955720d05a7c2d9c532 - v3.11-rc1 [SK - 0.7340]
Signed-off-by: Shubham Kulkarni <skulkarni@mvista.com>